find_distributors/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
gspread==6.1.4
google-auth==2.35.0
pydantic==2.11.7
